Extract slsk client class from index
